+++ src/TestSuite/TestTrait.php
@@ -17,6 +17,7 @@ declare(strict_types=1);
 namespace Tools\TestSuite;
 
 use BadMethodCallException;
+use Closure;
 use Exception;
 use PHPUnit\Framework\AssertionFailedError;
 use PHPUnit\Framework\Error\Deprecated;
@@ -266,6 +267,36 @@ trait TestTrait
         }
 
         return $this->getMockForAbstractClass($originalClassName, $arguments, '', true, true, true, $mockedMethods);
+    }
+
+    /**
+     * Helper method for check deprecation methods
+     * @param \Closure $callable callable function that will receive asserts
+     * @return void
+     * @since 1.8.0
+     * @codeCoverageIgnore
+     */
+    public function deprecated(Closure $callable): void
+    {
+        $previousHandler = set_error_handler(
+            function ($code, $message, $file, $line, $context = null) use (&$previousHandler, &$deprecation): bool {
+                if ($code == E_USER_DEPRECATED) {
+                    $deprecation = true;
+
+                    return true;
+                }
+                if ($previousHandler) {
+                    return $previousHandler($code, $message, $file, $line, $context);
+                }
+
+                return false;
+            }
+        );
+        try {
+            $callable();
+        } finally {
+        }
+        $this->assertTrue($deprecation, 'Should have at least one deprecation warning');
     }
 
     /**