+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 1.8 branch
 ### 1.8.1
 * tests for `ReflectionTrait` and `TestTrait` have largely improved;
+* the `TestCase::deprecated()` method has been moved to `TestTrait`;
 * the `Exceptionist` class is deprecated and will be removed in a later release;
 * all exception classes provided by this package are deprecated and will be removed in a later release;
 * fixed configuration of phpcs.